Magellan sim (#58)
* Starting work on Sim

* Added some basic logic

This commit just fleshes out a few more details, as of right now no ability to control rate of updates, will probably need Rate to included

* Added some logic to the sim

* Some adjustments

* added launch files

* moved files

* edited cmakelist (does not compile)

* Added info prints into console

* Last minute cleanup

* Lint changes

* Added a header file in include to see if CI will be happy

* Add q key (#59)

* Add ssh key

*  Removed extra blank line

* Changed velocity udpates to use math

* Changed velocity update logic

The sim now updates the velocity using simple kinematics. The updated values do currently oscillate if it is unable to match the velocity given the acceleration

* Ran formattit.sh

* Starting work on Sim

* Added some basic logic

This commit just fleshes out a few more details, as of right now no ability to control rate of updates, will probably need Rate to included

* Some adjustments

* added launch files

* reorganized files

* Added info prints into console

* Last minute cleanup and linting

* Changed velocity udpates to use math

* Changed velocity update logic

The sim now updates the velocity using simple kinematics. The updated values do currently oscillate if it is unable to match the velocity given the acceleration

* Ran formattit.sh